Major refactor of Parameters, and removal of TypeEditor files
