+++ src/client/src/compLibrary/dropdown/mimir/components/DropdownList.tsx
@@ -35,18 +35,16 @@ export const DropdownList = ({
   <DropdownListBox borderRadius={borderRadius} borderColor={borderColor} top={listTop}>
     {items?.map((item) => {
       return (
-        <>
-          <DropdownListItem
-            fontSize={fontSize}
-            height={height}
-            borderRadius={borderRadius}
-            onClick={() => handleChange(item)}
-            key={item[keyProp]}
-          >
-            {valueImageProp && <Symbol source={item[valueImageProp]} text={item[valueProp]} />}
-            <p>{item.name ?? item.key}</p>
-          </DropdownListItem>
-        </>
+        <DropdownListItem
+          fontSize={fontSize}
+          height={height}
+          borderRadius={borderRadius}
+          onClick={() => handleChange(item)}
+          key={item[keyProp]}
+        >
+          {valueImageProp && <Symbol source={item[valueImageProp]} text={item[valueProp]} />}
+          <p>{item.name ?? item.key}</p>
+        </DropdownListItem>
       );
     })}
   </DropdownListBox>
