--- src/client/src/models/data/typeEditor/Rds.ts
@@ -1,5 +1,0 @@
-export interface Rds {
-  id: string;
-  name: string;
-  iri: string;
-}
