--- src/client/src/models/data/typeEditor/SimpleType.ts
@@ -1,8 +1,0 @@
-import { AttributeType } from "./AttributeType";
-
-export interface SimpleType {
-  id: string;
-  name: string;
-  semanticReference: string;
-  attributes: AttributeType[];
-}
