--- src/client/src/models/data/typeEditor/TerminalType.ts
@@ -1,10 +1,0 @@
-import { AttributeType } from "./AttributeType";
-
-export interface TerminalType {
-  id: string;
-  name: string;
-  color: string;
-  terminalCategoryId: string;
-  semanticReference: string;
-  attributes: AttributeType[];
-}
