--- src/client/src/models/data/typeEditor/TerminalTypeDict.ts
@@ -1,3 +1,0 @@
-import { TerminalTypeDictItem } from "./TerminalTypeDictItem";
-
-export type TerminalTypeDict = TerminalTypeDictItem[];
