--- src/client/src/models/data/typeEditor/TerminalTypeDictItem.ts
@@ -1,6 +1,0 @@
-import { TerminalType } from "./TerminalType";
-
-export interface TerminalTypeDictItem {
-  key: string;
-  value: TerminalType[];
-}
