--- src/client/src/models/data/typeEditor/TerminalTypeExtended.ts
@@ -1,8 +1,0 @@
-import { ConnectorDirection } from "@mimirorg/modelbuilder-types";
-import { TerminalType } from "./TerminalType";
-
-export interface TerminalTypeExtended extends TerminalType {
-  type: ConnectorDirection;
-  terminalTypeId: string;
-  number: number;
-}
