--- src/client/src/models/data/typeEditor/TerminalTypeItem.ts
@@ -1,10 +1,0 @@
-import { ConnectorDirection } from "@mimirorg/modelbuilder-types";
-
-export interface TerminalTypeItem {
-  terminalId: string;
-  terminalTypeId: string;
-  selected: boolean;
-  connectorType: ConnectorDirection;
-  number: number;
-  categoryId: string;
-}
