+++ src/client/src/models/index.ts
@@ -23,15 +23,3 @@ export type { ProjectFileAm } from "./application/ProjectFileAm";
 export type { CombinedAttribute } from "./application/CombinedAttribute";
 export type { LibrarySubProjectItem } from "./application/LibrarySubProjectItem";
 export type { CombinedAttributeFilter } from "./application/CombinedAttributeFilter";
-
-//Type Editor
-export type { Rds } from "./data/typeEditor/Rds";
-export type { SimpleType } from "./data/typeEditor/SimpleType";
-export type { TerminalType } from "./data/typeEditor/TerminalType";
-export type { AttributeType } from "./data/typeEditor/AttributeType";
-export type { TerminalTypeDict } from "./data/typeEditor/TerminalTypeDict";
-export type { TerminalTypeItem } from "./data/typeEditor/TerminalTypeItem";
-export type { TerminalTypeExtended } from "./data/typeEditor/TerminalTypeExtended";
-export type { TerminalTypeDictItem } from "./data/typeEditor/TerminalTypeDictItem";
-
-export type { LocationType } from "./application/typeEditor/LocationType";
