--- src/client/src/modules/inspector/components/tabs/components/terminals/components/selector/components/helpers/FormatTypeId.ts
@@ -1,4 +1,0 @@
-import { ConnectorDirection } from "@mimirorg/modelbuilder-types";
-import { TerminalType } from "../../../../../../../../../../models";
-
-export const FormatTypeId = (type: TerminalType, connectorType: ConnectorDirection) => `${type.id} ${connectorType}`;
