+++ src/client/src/redux/sagas/index.ts
@@ -44,25 +44,6 @@ import {
   searchProject,
   updateProject,
 } from "./project/saga";
-// import {
-//   getAttributes,
-//   getBlobData,
-//   getInitialData,
-//   getLocationTypes,
-//   getPredefinedAttributes,
-//   getRDS,
-//   getSelectedCreateLibraryType,
-//   getSimpleTypes,
-//   getTerminals,
-//   saveType,
-// } from "./typeEditor/saga";
-// import {
-//   fetchBlobData,
-//   fetchCreateLibraryType,
-//   fetchInitialData,
-//   fetchSimpleTypes,
-//   saveLibraryType,
-// } from "../../typeEditor/redux/typeEditorSlice";
 
 //TODO: Add takeEvery for LOCK_ on
 function* sagas() {
@@ -83,16 +64,6 @@ function* sagas() {
     takeEvery(fetchLibraryTerminals, getTerminals),
     takeEvery(exportLibraryAction, exportLibrary),
     takeEvery(importLibraryAction, importLibrary),
-    // takeEvery(fetchInitialData, getInitialData),
-    // takeEvery(fetchInitialData, getLocationTypes),
-    // takeEvery(fetchInitialData, getRDS),
-    // takeEvery(fetchInitialData, getTerminals),
-    // takeEvery(fetchInitialData, getPredefinedAttributes),
-    // takeEvery(fetchInitialData, getAttributes),
-    // takeEvery(fetchSimpleTypes, getSimpleTypes),
-    // takeEvery(fetchBlobData, getBlobData),
-    // takeEvery(fetchCreateLibraryType, getSelectedCreateLibraryType),
-    // takeEvery(saveLibraryType, saveType),
     takeEvery(deleteLibraryAction, deleteLibraryItem),
   ]);
 }
